feat(mind-map): let the node panel set text, background and branch colors
Extracts the color picker out of NoteColorPicker into widgets/react so both
the note-bound variant and the mind map share it, and gives it two things the
mind map needs: overridable tooltips (the existing strings are worded for
notes) and an indeterminate state for a selection whose nodes disagree.

The panel edits every selected node at once. Clearing writes an empty string
rather than dropping the property, because Mind Elixir only ever assigns the
style properties it is handed and never resets the others.
