+++ apps/client/src/widgets/react/custom-color.png
